requirements.txt
Django==3.2
djangorestframework==3.12.4
drf-spectacular==0.15.1
factory-boy==3.2.0
gunicorn==20.1.0
psycopg2-binary==2.8.6
Scrapy==2.5.0
